+++ .gitignore
@@ -1,13 +1,2 @@
-*.pyc
-*~
-.*.sw[op]
-*.egg-info
-*.class
-*.o
-.coverage
-dist
-build
-_build
-distribute-*
-local
-.tox
+/.externalNativeBuild
+/build
