--- CONTRIBUTING.md
@@ -1,8 +1,0 @@
-# Contributing
-
-PyBee <3's contributions! 
-
-Please be aware, PyBee operates under a Code of Conduct. 
-
-See [CONTRIBUTING to PyBee](http://pybee.org/contributing) for details.
-
