+++ build.gradle
@@ -1,0 +1,57 @@
+import com.android.build.OutputFile
+
+apply plugin: 'com.android.library'
+
+android {
+    compileSdkVersion 28
+
+
+    defaultConfig {
+        minSdkVersion 22
+        targetSdkVersion 28
+        versionCode 1
+        versionName "1.0"
+        ndk {
+            // abiFilters "x86", "x86_64", "armeabi-v7a", "arm64-v8a"
+            abiFilters "armeabi-v7a"
+        }
+        externalNativeBuild {
+            cmake {
+                cppFlags ""
+            }
+        }
+    }
+
+    buildTypes {
+        release {
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
+        }
+    }
+
+    sourceSets {
+        main {
+            // Include the Python native support libraries
+            jniLibs.srcDirs = [
+                    "../python/lib",
+            ]
+            // Add the Python home as assets
+            assets.srcDirs += '../python/home'
+            // Add the Python app code as an asset
+            assets.srcDirs += 'build/python'
+        }
+    }
+
+    externalNativeBuild {
+        cmake {
+            version '3.10.2'
+            path "src/main/cpp/CMakeLists.txt"
+        }
+    }
+
+}
+
+dependencies {
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    implementation 'com.android.support:appcompat-v7:28.0.0'
+}
