--- org/pybee/rubicon/test/ICallback.java
@@ -1,8 +1,0 @@
-package org.pybee.rubicon.test;
-
-
-public interface ICallback {
-    public void poke(Example example, int value);
-
-    public void peek(Example example, int value);
-}
