--- setup.py
@@ -1,45 +1,0 @@
-#/usr/bin/env python
-import io
-import re
-from setuptools import setup, find_packages
-
-
-with io.open('./rubicon/java/__init__.py', encoding='utf8') as version_file:
-    version_match = re.search(r"^__version__ = ['\"]([^'\"]*)['\"]", version_file.read(), re.M)
-    if version_match:
-        version = version_match.group(1)
-    else:
-        raise RuntimeError("Unable to find version string.")
-
-
-with io.open('README.rst', encoding='utf8') as readme:
-    long_description = readme.read()
-
-
-setup(
-    name='rubicon-java',
-    version=version,
-    description='A bridge interface between Python and Java.',
-    long_description=long_description,
-    author='Russell Keith-Magee',
-    author_email='russell@keith-magee.com',
-    url='http://pybee.org/rubicon',
-    packages=find_packages(exclude=['tests']),
-    namespace_packages=['rubicon'],
-    license='New BSD',
-    classifiers=[
-        'Development Status :: 3 - Alpha',
-        'Intended Audience :: Developers',
-        'License :: OSI Approved :: BSD License',
-        'Programming Language :: Java',
-        'Programming Language :: Python :: 2',
-        # 'Programming Language :: Python :: 2.6',
-        'Programming Language :: Python :: 2.7',
-        # 'Programming Language :: Python :: 3',
-        # 'Programming Language :: Python :: 3.3',
-        # 'Programming Language :: Python :: 3.4',
-        'Topic :: Software Development',
-        'Topic :: Software Development :: User Interfaces',
-        'Topic :: Software Development :: Widget Sets',
-    ],
-)
