+++ src/main/cpp/python_log_helper.h
@@ -1,0 +1,5 @@
+#pragma once
+
+start_redirecting_stdout_stderr();
+stop_log_redirect();
+
