+++ src/main/java/org/beeware/rubicon/test/ICallback.java
@@ -1,0 +1,8 @@
+package org.beeware.rubicon.test;
+
+
+public interface ICallback {
+    public void poke(Example example, int value);
+
+    public void peek(Example example, int value);
+}
