+++ src/main/python/app_packages/rubicon/java/__init__.py
@@ -1,0 +1,5 @@
+from .api import *
+from .jni import *
+from .types import *
+
+__version__ = '0.1.0'
