--- tests/runner.py
@@ -1,4 +1,0 @@
-from unittest.main import TestProgram
-
-if __name__ == '__main__':
-    TestProgram(argv=[__file__,'discover'], verbosity=2)
