--- tox.ini
@@ -1,15 +1,0 @@
-# Tox (http://tox.testrun.org/) is a tool for running tests
-# in multiple virtualenvs. This configuration file will run the
-# test suite on all supported python versions. To use it, "pip install tox"
-# and then run "tox" from this directory.
-
-[tox]
-envlist = py26, py27, py33, py34
-
-[testenv]
-commands = {envpython} setup.py test
-deps =
-
-[testenv:py26]
-deps =
-    unittest2
